text_summarizer/requirements.txt
kafka-python==2.0.2
sumy==0.11.0
numpy==1.24.2